+++ DesignToolbox/DesignToolbox/Pages/Components/ComponentsPage.swift
@@ -27,6 +27,7 @@ struct ComponentsPage: View {
         LinkElement(),
         RadioElements(),
         SwitchElements(),
+        TagElement(),
     ]
 
     var body: some View {